Some Changes

+++ app/src/main/java/com/webtalented/addquestion/Delete.kt
@@ -4,12 +4,10 @@ import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import kotlinx.android.synthetic.main.activity_delete.*
-import android.support.annotation.NonNull
-import android.support.v4.app.FragmentActivity
 import android.util.Log
 import android.widget.Toast
-import com.google.android.gms.tasks.OnFailureListener
-import com.google.android.gms.tasks.OnSuccessListener
+
+
 import com.google.firebase.firestore.FirebaseFirestore
 
 
@@ -22,15 +20,21 @@ class Delete : AppCompatActivity() {
         val db = FirebaseFirestore.getInstance()
 
         sil.setOnClickListener { it ->
-            db.collection("Suallar").document(id.text.toString())
-                    .delete()
-                    .addOnSuccessListener {
-                        val intent = Intent(this , MainActivity::class.java)
-                        startActivity(intent)
-                        Log.d(this@Delete.toString(), "DocumentSnapshot successfully deleted!") }
-                    .addOnFailureListener { e ->
-                        Toast.makeText(this@Delete, "Silinmedi" , Toast.LENGTH_LONG).show()
-                        Log.w(this@Delete.toString(), "Error deleting document", e) }
+            if(id.text.isNotEmpty()){
+
+                db.collection("Suallar").document(id.text.toString())
+                        .delete()
+                        .addOnSuccessListener {
+                            val intent = Intent(this , MainActivity::class.java)
+                            startActivity(intent)
+                            Log.d(this@Delete.toString(), "DocumentSnapshot successfully deleted!") }
+                        .addOnFailureListener { e ->
+                            Toast.makeText(this@Delete, "Silinmedi" , Toast.LENGTH_LONG).show()
+                            Log.w(this@Delete.toString(), "Error deleting document", e) }
+
+            }else{
+                Toast.makeText(this@Delete, "Id daxil edilmedi", Toast.LENGTH_LONG).show()
+            }
         }
     }
 }
